Omg, a commit in english!

+++ src/routes/exchanger/components/ItemCard.tsx
@@ -8,23 +8,30 @@ import { endPoints, routes, serverAddress } from 'src/utils/constants';
 import RoutesHandler from 'src/utils/routesHandler';
 
 import { FaRegTrashAlt, FaEdit } from 'react-icons/fa';
+import EditItemModal from './inventory/EditItemModal';
+import ConfirmationModal from 'src/components/modals/Confirmation';
+import { useState } from 'react';
 //la publicacion particular del producto a intercambiar
 
-export default function ItemCard({ item, onClick, hiddeOwner, canDelete, queryInvalidator, isEditable }: ItemCardProps) {
+export default function ItemCard({ item, onClick, hiddeOwner, queryInvalidator, hiddeBtns, isEditable }: ItemCardProps) {
   
-  const { photo, name, description, owner, itemCategory, quantity } = item
-  const { closeModal } = useCustomModal()
+  const [ itemData, setItemData ] = useState(item)
+  const { showModal, closeModal } = useCustomModal()
   const { setRoute } = RoutesHandler()
 
   const onClickOwner = (ev: MouseEvent) => {
     ev.stopPropagation()
     closeModal()
-    setRoute(`${routes.exchanger.profile}/${owner?.id}`)
+    setRoute(`${routes.exchanger.profile}/${itemData.owner?.id}`)
   }
 
   const onClickExchange = (ev: MouseEvent) => {
     ev.stopPropagation()
     console.log('Intercambiar')
+  }
+
+  const onClickEdit = () => {
+    showModal(<EditItemModal itemData={item} onEditItem={(data) => setItemData({ ...itemData, ...data })} />)
   }
 
   const { mutate: deleteItem } = useMutation({
@@ -37,33 +44,31 @@ export default function ItemCard({ item, onClick, hiddeOwner, canDelete, queryIn
     }
   })
 
+  const confirmation = (fn) => showModal(<ConfirmationModal onAccept={fn}/>)
+
   return (
     <div className="bg-white p-4 rounded shadow-lg max-w-sm cursor-pointer transform transition-transform duration-200 hover:scale-105" onClick={onClick}>
-      <Image photo={photo} alt={name} className="mb-4 w-full rounded shadow-2xl" />
-      <h2 className="text-xl font-bold mb-2">{name}</h2>
-      <p className="text-sm text-gray-500 mb-2">Categoria: {itemCategory.name}</p>
-      <p className="text-sm text-gray-500 mb-2">Descripcion: {description}</p>
+      <Image photo={itemData.photo} alt={itemData.name} className="mb-4 w-full rounded shadow-2xl" />
+      <h2 className="text-xl font-bold mb-2">{itemData.name}</h2>
+      <p className="text-sm text-gray-500 mb-2">Categoria: {itemData.itemCategory.name}</p>
+      <p className="text-sm text-gray-500 mb-2">Descripcion: {itemData.description}</p>
       <div className='flex items-center justify-between'>
-        <p className="text-sm text-gray-500 mb-2">Cantidad restante: {quantity}</p>
-        <div className='flex gap-2'>
-          {
-            isEditable &&
-            <button className='btn'><FaEdit/></button>
-          }
-          {
-            canDelete &&
-            <button onClick={() => deleteItem()} className='btn btn-error'><FaRegTrashAlt/></button>
-          }
-        </div>
+        <p className="text-sm text-gray-500 mb-2">Cantidad restante: {itemData.quantity}</p>
+        {
+          isEditable && <div className='flex gap-2'>
+            <button className='btn' onClick={onClickEdit}><FaEdit/></button>
+            <button onClick={() => confirmation(deleteItem)} className='btn btn-error'><FaRegTrashAlt/></button>
+          </div>
+        }
       </div>
       {
-        !canDelete &&
+        !hiddeBtns && 
         <div className="flex flex-col items-start">
             <button onClick={onClickExchange} className="bg-red-500 text-white px-4 py-2 rounded mb-2 transform transition-transform duration-200 hover:scale-105">
               Intercambiar
             </button>
             <button onClick={onClickOwner} className={`bg-blue-500 ${hiddeOwner && 'hidden'} text-white px-4 py-2 rounded transform transition-transform duration-200 hover:scale-105`}>
-              Dueño del item: {owner.name}
+              Dueño del item: {itemData.owner.name}
             </button>
         </div>
       }
